Added tool to plot the timeseries data
Added a simple tool to read the full timeseries
data and emit selected data (timestamp, outstanding-bytes).

Added a python program that uses the emitted data
and plot it. It uses matplotlib.

Updated README with instructions for plotting
and a complete example.

Fix code for the outstanding computation
and the associated test.